\name{AssociatedPress}
\alias{AssociatedPress}
\docType{data}
\title{Associated Press data}
\description{
  Associated Press data from the First Text Retrieval Conference
  (TREC-1) 1992.
}
\usage{data("AssociatedPress")}
\format{
  The data set is an object of class \code{"DocumentTermMatrix"}
  provided by package \pkg{tm}. It is a document-term matrix which
  contains the term frequency of 10473 terms in 2246 documents.
}
\source{
  Accompanying material to the source code for fitting LDA models
  provided by David M. Blei and co-authors. Downloaded from:
  \url{http://www.cs.berkeley.edu/~blei}.
}
\references{
  D. Harman (1992) Overview of the first text retrieval conference
  (TREC-1). In Proceedings of the First Text Retrieval Conference
  (TREC-1), 1--20.
}
\keyword{datasets}
